Fix redundant OnHMIStatus notification
This issue is happening when temporary events
are applied. In this case OnHMIStatus notification
sending is triggered from 2 places: after changing
of audio streaming state of the main window and
after applying temp state to the actual window.
As due to proposal all temporary events are applied
to all application windows, there is no need to
track audio streaming state change in a separate
function so it was removed.
Also, in temp events was added check for the window
type to prevent wrong HMI level calculation for
widgets.